OFI: EP_MSG implemented
This finishes the implementation of the support for connected enpoint in OFI
This allows to set the type of endpoint in the config to FI_EP_RDM or FI_EP_MSG
to support more launch configurations

+++ src/MPC_Common/include/mpc_mempool.h
@@ -45,7 +45,7 @@ void _mpc_mempool_stack(void * buf);
 int mpc_mempool_init(mpc_mempool *mp, 
     int min, 
     int max, 
-    int size, 
+    size_t size, 
     void *(*malloc_func)(size_t), 
     void (*free_func)(void *));
